Add user autocomplete to bennu toolkit.
To use autocomplete for an input field add bennu-user-autocomplete attribute when using vanilla javacript.
The original input will have the user's id after selecting a user.

If using angular-js use the ng-user-autocomplete directive. The value of the ng-user-autocomplete is the user model used to store the selected user.
There is an usage example (in angularjs) in the file src/main/webapp/bennu-admin/template/Groups.html