Bennu Scheduler now allows for Future-like submission of arbitrary code to be executed sometime later with PersistentFuture. (#322)
Standard Java Serialization rules apply - Runnable class, or the class where lambda is defined (when using lambda), must implement Serializable.
Each submission is only executed once. In case the code is no longer deserializable when atempting to execute it, or if for any other reason it throws an Exception, the code is considered to have been executed, even if unsuccessfully.