Login page uses locale from browser
Instead of providing the login page with the application's default
locale we first try to match the browsers' locale with one of the
application's supported locales, if there's a match we server the login
page in that locale, otherwise we'll use the application's default
locale.

Note: if the user has a different locale selected in their userProfile
after login that setting will supersede any browser locale information.