Bug fix: Correct use of streams in GenericFile
The CountingStream was always returning zero size because the stream
will only be consumed after the TX checkpoint (since the
LocalFileSystemStorage creates a FileWriteIntention in a PerTxBox).

Besides that issue, the tika detector was consuming the first N bytes of
the stream creating issues as well.

With the current commit we fully consume the stream to detect the
content type and have the CountingInputStream return the correct size.
After that we reset the stream and pass it to the file storage.