Transforming logout filter into a servlet
Being a filter if any filter performs a forward to "/logout" it will
result in a 404, since a forward will skip the rest of the filters.
We were seeing this sort of behaviour in the SAML filter processing
logout messages, transforming this filter into a servlet allows that
filters can directly just to the dispatcher and still be able to perform
the logout operation.

Relates with: #qubIT-Omnis-1882