Removing InjectCheckSumInUrlFunction from Portal
This would create a ciclic dependency since when we compile bennu we
compile bennu-core and bennu-portal as one, and bennu-portal depended
in bennu-renderers and bennu-renderers depended in bennu-core.

We could have introduced an interface to break this dependency chaining
but after reviewing the code we identified that the function is no longer
used in any templates and GenericChecksumRewriter is being used directly
in JSPs when needed.

So it's best to just remove the code.

Relates with: #qubIT-Omnis-5780